feat: Introduce dynamic rendering for resume and cover letter components with section registries
- Added new section registries for both resume and cover letter components to facilitate dynamic rendering based on visibility and data presence.
- Implemented individual components for each section, enhancing modularity and maintainability.
- Updated the resume and cover letter schemas to support the new structure, ensuring consistency across templates.
- Enhanced unit tests to validate the visibility logic and rendering behavior of the new components, improving overall reliability.
- Removed outdated section registry files to streamline the codebase and reduce redundancy.
