+++ resume-data/tailor/tech-corp/resume.yaml
@@ -10,7 +10,6 @@ resume:
   summary: 'Passionate Senior Frontend Engineer with a decade of experience building scalable applications serving millions of users daily. Expert in React, TypeScript, and modern frontend architecture, with proven impact delivering APIs handling 1M+ requests/day and leading platform migrations that improved performance by 30%. Dedicated to best practices, component-driven development, and exceptional user experience across cross-functional teams.'
 
   contact:
-    name: 'John Doe'
     phone: '+1 (555) 555-5555'
     email: 'john.doe@example.com'
     address: '456 Innovation Drive, San Francisco, CA 94105'
