+++ src/templates/modern/components/resume/Education.tsx
