--- src/templates/modern/resume/components/List.tsx
@@ -1,29 +1,0 @@
-import React from 'react';
-import { Text, View, StyleSheet } from '@react-pdf/renderer';
-
-const styles = StyleSheet.create({
-  item: {
-    flexDirection: 'row',
-    marginBottom: 5,
-  },
-  bulletPoint: {
-    width: 10,
-    fontSize: 10,
-  },
-  itemContent: {
-    flex: 1,
-    fontSize: 10,
-    fontFamily: 'Lato',
-  },
-});
-
-const List = ({ children }: { children: React.ReactNode }) => children;
-
-export const Item = ({ children }: { children: React.ReactNode }) => (
-  <View style={styles.item}>
-    <Text style={styles.bulletPoint}>•</Text>
-    <Text style={styles.itemContent}>{children}</Text>
-  </View>
-);
-
-export default List;
