--- src/templates/modern/resume/components/Title.tsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import { Text, StyleSheet } from '@react-pdf/renderer';
-
-const styles = StyleSheet.create({
-  title: {
-    fontFamily: 'Lato Bold',
-    fontSize: 14,
-    marginBottom: 10,
-    textTransform: 'uppercase',
-  },
-});
-
-const Title = ({ children }: { children: React.ReactNode }) => (
-  <Text style={styles.title}>{children}</Text>
-);
-
-export default Title;
